Rename utils and convert to TS
